Restructure the repo as a workspace (#233)
* Add a workspace

* go

* Fix tests concurrency

* Update websocket proxy location

* Fix

* Fix

* Update cargo run commands

* Use workspace dependencies

* Fix issues

* update deny

* More deny stuff
