+++ .github/workflows/websocket-proxy-ci.yaml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: "./websocket-proxy"
+        working-directory: "./crates/websocket-proxy"
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
