+++ README.md
@@ -7,7 +7,7 @@ Rollup Boost is a block builder sidecar for Optimism Stack chains to enable exte
 Run the rollup-boost server using the following command:
 
 ```
-cargo run -- [OPTIONS]
+cargo run --bin rollup-boost -- [OPTIONS]
 ```
 
 ### Command-line Options
@@ -35,7 +35,7 @@ You can also set the options using environment variables. See .env.example to us
 ### Example
 
 ```
-cargo run --l2-jwt-token your_jwt_token --l2-url http://localhost:8545 --builder-jwt-token your_jwt_token --builder-url http://localhost:8546
+cargo run --bin rollup-boost -- --l2-jwt-token your_jwt_token --l2-url http://localhost:8545 --builder-jwt-token your_jwt_token --builder-url http://localhost:8546
 ```
 
 ## Core System Workflow
