+++ crates/rollup-boost/Cargo.toml
@@ -1,0 +1,73 @@
+[package]
+name = "rollup-boost"
+version = "0.1.0"
+edition = "2024"
+license = "MIT"
+
+[dependencies]
+tracing.workspace = true
+tracing-subscriber.workspace = true
+clap.workspace = true
+futures.workspace = true
+thiserror.workspace = true
+serde_json.workspace = true
+metrics.workspace = true
+tokio.workspace = true
+
+op-alloy-rpc-types-engine = "0.12.0"
+alloy-rpc-types-engine = "0.13.0"
+alloy-rpc-types-eth = "0.13.0"
+alloy-primitives = { version = "0.8.10", features = ["rand"] }
+serde = { version = "1", features = ["derive"] }
+jsonrpsee = { version = "0.24", features = ["server", "http-client", "macros"] }
+moka = { version = "0.12.10", features = ["future"] }
+http = "1.1.0"
+dotenvy = "0.15.7"
+tower = { version = "0.4.13", features = ["timeout"] }
+tower-http = { version = "0.5.2", features = [
+    "decompression-full",
+    "sensitive-headers",
+] }
+http-body-util = "0.1.2"
+hyper = { version = "1.4.1", features = ["full"] }
+hyper-util = { version = "0.1", features = ["full"] }
+hyper-rustls = { version = "0.27.0", features = ["ring"] }
+rustls = { version = "0.23.23", features = ["ring"] }
+opentelemetry = { version = "0.28.0", features = ["trace"] }
+opentelemetry-otlp = { version = "0.28.0", features = [
+    "http-proto",
+    "http-json",
+    "reqwest-client",
+    "trace",
+    "grpc-tonic",
+] }
+opentelemetry_sdk = { version = "0.28.0", features = ["rt-tokio"] }
+tracing-opentelemetry = "0.29.0"
+metrics-exporter-prometheus = "0.16.0"
+metrics-util = "0.19.0"
+eyre = "0.6.12"
+paste = "1.0.15"
+parking_lot = "0.12.3"
+
+[dev-dependencies]
+rand = "0.9.0"
+time = { version = "0.3.36", features = ["macros", "formatting", "parsing"] }
+op-alloy-consensus = "0.12.0"
+alloy-eips = { version = "0.13.0", features = ["serde"] }
+alloy-consensus = { version = "0.13.0", features = ["serde"] }
+anyhow = "1.0"
+testcontainers = { version = "0.23.3" }
+assert_cmd = "2.0.10"
+predicates = "3.1.2"
+tokio-util = { version = "0.7.13" }
+bytes = "1.2"
+reth-rpc-layer = { git = "https://github.com/paradigmxyz/reth.git", rev = "v1.3.7" }
+ctor = "0.4.1"
+reqwest = "0.12.15"
+
+[[bin]]
+name = "rollup-boost"
+path = "src/bin/main.rs"
+
+[lib]
+path = "src/lib.rs"
