+++ crates/rollup-boost/src/bin/main.rs
@@ -1,10 +1,14 @@
 use clap::Parser;
 use rollup_boost::Args;
+use rollup_boost::init_tracing;
 
 use dotenvy::dotenv;
 
 #[tokio::main]
 async fn main() -> eyre::Result<()> {
     dotenv().ok();
-    Args::parse().run().await
+
+    let args = Args::parse();
+    init_tracing(&args)?;
+    args.run().await
 }
