+++ crates/rollup-boost/src/cli.rs
@@ -12,7 +12,7 @@ use crate::{
     BlockSelectionPolicy, DebugClient, ProxyLayer, RollupBoostServer, RpcClient,
     client::rpc::{BuilderArgs, L2ClientArgs},
     debug_api::ExecutionMode,
-    init_metrics, init_tracing,
+    init_metrics,
     payload::PayloadSource,
     probe::ProbeLayer,
 };
@@ -122,7 +122,6 @@ impl Args {
             };
         }
 
-        init_tracing(&self)?;
         init_metrics(&self)?;
 
         let l2_client_args = self.l2_client;
