+++ crates/rollup-boost/src/client/auth.rs
