+++ crates/rollup-boost/src/debug_api.rs
