+++ crates/rollup-boost/src/selection.rs
