+++ crates/rollup-boost/src/tests/common/mod.rs
@@ -1,4 +1,8 @@
 #![allow(dead_code)]
+use crate::DebugClient;
+use crate::{AuthLayer, AuthService};
+use crate::{EngineApiClient, OpExecutionPayloadEnvelope, PayloadVersion};
+use crate::{NewPayload, PayloadSource};
 use alloy_eips::Encodable2718;
 use alloy_primitives::{B256, Bytes, TxKind, U256, address, hex};
 use alloy_rpc_types_engine::{ExecutionPayload, JwtSecret};
@@ -17,10 +21,6 @@ use op_alloy_consensus::TxDeposit;
 use op_alloy_rpc_types_engine::OpPayloadAttributes;
 use parking_lot::Mutex;
 use proxy::{BuilderProxyHandler, start_proxy_server};
-use rollup_boost::DebugClient;
-use rollup_boost::EngineApiClient;
-use rollup_boost::{AuthLayer, AuthService};
-use rollup_boost::{NewPayload, OpExecutionPayloadEnvelope, PayloadSource, PayloadVersion};
 use serde_json::Value;
 use services::op_reth::{AUTH_RPC_PORT, OpRethConfig, OpRethImage, OpRethMehods, P2P_PORT};
 use services::rollup_boost::{RollupBoost, RollupBoostConfig};
@@ -45,7 +45,7 @@ use tower_http::sensitive_headers::SetSensitiveRequestHeaders;
 pub const JWT_SECRET: &str = "688f5d737bad920bdfb2fc2f488d6b6209eebda1dae949a8de91398d932c517a";
 pub const L2_P2P_ENODE: &str = "3479db4d9217fb5d7a8ed4d61ac36e120b05d36c2eefb795dc42ff2e971f251a2315f5649ea1833271e020b9adc98d5db9973c7ed92d6b2f1f2223088c3d852f";
 pub static TEST_DATA: LazyLock<String> =
-    LazyLock::new(|| format!("{}/tests/common/test_data", env!("CARGO_MANIFEST_DIR")));
+    LazyLock::new(|| format!("{}/src/tests/common/test_data", env!("CARGO_MANIFEST_DIR")));
 
 pub mod proxy;
 pub mod services;
@@ -256,7 +256,7 @@ impl RollupBoostTestHarnessBuilder {
         let timestamp = dt.format(&format)?;
 
         let dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("integration_logs")
+            .join("../../integration_logs")
             .join(self.test_name.clone())
             .join(timestamp);
         std::fs::create_dir_all(&dir)?;
