+++ crates/rollup-boost/src/tests/common/services/op_reth.rs
@@ -5,7 +5,7 @@ use testcontainers::{
     core::{ContainerPort, WaitFor},
 };
 
-use crate::common::TEST_DATA;
+use crate::tests::common::TEST_DATA;
 
 const NAME: &str = "ghcr.io/paradigmxyz/op-reth";
 const TAG: &str = "v1.3.12";
