+++ crates/rollup-boost/src/tests/common/test_data/jwt_secret.hex
