+++ crates/rollup-boost/src/tests/fcu_no_block_time_delay.rs
@@ -1,7 +1,5 @@
-mod common;
+use super::common::{RollupBoostTestHarnessBuilder, proxy::BuilderProxyHandler};
 
-use common::RollupBoostTestHarnessBuilder;
-use common::proxy::BuilderProxyHandler;
 use futures::FutureExt;
 use serde_json::Value;
 use std::pin::Pin;
