+++ crates/rollup-boost/src/tests/mod.rs
@@ -1,0 +1,11 @@
+mod common;
+
+mod builder_full_delay;
+mod builder_returns_incorrect_block;
+mod execution_mode;
+mod fcu_no_block_time_delay;
+mod no_tx_pool;
+mod remote_builder_down;
+mod simple;
+mod simple_isthmus;
+mod simple_isthmus_transition;
