+++ crates/rollup-boost/src/tests/no_tx_pool.rs
@@ -1,6 +1,4 @@
-mod common;
-
-use crate::common::RollupBoostTestHarnessBuilder;
+use super::common::RollupBoostTestHarnessBuilder;
 
 #[tokio::test]
 async fn no_tx_pool() -> eyre::Result<()> {
