+++ crates/rollup-boost/src/tracing.rs
