+++ crates/websocket-proxy/.dockerignore
