+++ crates/websocket-proxy/Cargo.toml
@@ -3,30 +3,35 @@ name = "flashblocks-websocket-proxy"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.85"
+license = "MIT"
 
 [dependencies]
-tokio = { version = "1.44.2", features = ["full"] }
+tracing.workspace = true
+tracing-subscriber.workspace = true
+clap.workspace = true
+futures.workspace = true
+thiserror.workspace = true
+serde_json.workspace = true
+metrics.workspace = true
+tokio.workspace = true
+
 tokio-tungstenite = { version = "0.26.2", features = ["native-tls"] }
-metrics-exporter-prometheus = { version = "0.17.0", features = ["http-listener"]}
+metrics-exporter-prometheus = { version = "0.17.0", features = [
+    "http-listener",
+] }
 http = "1.2.0"
 axum = { version = "0.8.1", features = ["ws"] }
-tracing = "0.1.41"
-clap = { version = "4", features = ["derive", "env"] }
 dotenvy = "0.15.7"
 backoff = "0.4.0"
-futures = "0.3.31"
-tracing-subscriber = { version = "0.3.19", features = ["json", "env-filter"] }
-tokio-util = "0.7.12"
-reqwest = { version = "0.12.15", default-features = false, features = ["native-tls"] }
-metrics = "0.24.1"
+reqwest = { version = "0.12.15", default-features = false, features = [
+    "native-tls",
+] }
 metrics-derive = "0.1"
-thiserror = "2.0.11"
-serde_json = "1.0.138"
 hostname = "0.4.0"
 redis = "0.30.0"
 redis-test = { version = "0.10.0", optional = true }
 uuid = { version = "1.16.0", features = ["v4"] }
-
+tokio-util = "0.7.12"
 
 [dependencies.ring]
 version = "0.17.12"
