+++ crates/websocket-proxy/Dockerfile
