+++ crates/websocket-proxy/src/registry.rs
