+++ deny.toml
@@ -10,6 +10,7 @@ allow = [
   "BSD-3-Clause",
   "Zlib",
   "OpenSSL",
+  "CDLA-Permissive-2.0",
 ]
 confidence-threshold = 0.8
 
@@ -17,6 +18,10 @@ confidence-threshold = 0.8
 ignore = [
   # paste crate is unmaintained
   "RUSTSEC-2024-0436",
+  # backoff crate is unmaintained
+  "RUSTSEC-2025-0012",
+  # instant crate is unmaintained  
+  "RUSTSEC-2024-0384",
 ]
 
 [bans]
